requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
python-multipart==0.0.6